style: Add footer to App.vue
This commit adds a footer to the App.vue component in order to display a "Made with love" message at the bottom of the page. The footer is positioned fixed at the bottom of the screen and has a light gray background color with dark gray text. This change improves the overall user experience by adding a personal touch to the application.